Refactor, add comments, and update tests.
Rewrite trait to use prefixed property names instead of those in the DB, for increased clarity and specificity and so as not to interfere with `data_get()` and other array operations. Add PHPDoc comments to all methods on main `HasTranslations` trait. Install `codedungeon/phpunit-result-printer` for pretty test output. Update namespace and rename lots of things. Update all tests to reflect these changes, and ensure they pass.